Revert "GCW-2995 fixed multiple subscription via websocket"
This reverts commit bbbd3b177756b6554c28208d9cd38e6da82727d8.
